+++ app/controllers/anonymous_feedback/problem_reports/explore_controller.rb
@@ -1,7 +1,7 @@
 require 'support/requests/anonymous/problem_report'
 require 'support/requests/anonymous/explore'
 
-class AnonymousFeedback::ProblemReports::ExploreController < ApplicationController
+class AnonymousFeedback::ProblemReports::ExploreController < AuthorisationController
   authorize_resource class: Support::Requests::Anonymous::ProblemReport
 
   def new
